+++ dave/server/languages/c_cpp/__init__.py
@@ -1,3 +1,4 @@
 from .std_1D import *
 from .std_2D import *
 from .juce import *
+from .choc import *